+++ wiki/Changelog.md
@@ -6,6 +6,7 @@ Terminal features
   * Avoid repetitive painting of images.
   * Fix selection highlighting in configured colours on emojis.
   * Full or dimmed selection highlighting depending on clipboard contents (#1322).
+  * Fix GR-mapped character sets in locking shift mode (vttest 3.10.).
 
 WSL support
   * New mode to run WSL directly with Windows default access gateway wsl.exe.